fixed as : Decrease a surface from Produits, when i create a Bon Livraisons

+++ backend/models/BonLivraisonProduit.js
@@ -12,7 +12,7 @@ const BonLivraisonProduit = sequelize.define(
     },
 
     quantite: {
-      type: DataTypes.DECIMAL(10, 2),
+      type: DataTypes.INTEGER,
       allowNull: false,
       defaultValue: 1,
     },
